+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove local kubeconfig file to avoid issues when recreating cluster with same ID.
+
 ## [0.1.0] - 2021-09-14
 
 ### Added